requirements-dev.txt
cookiecutter
pytest-cookies
dash[testing]
black
pre-commit
gunicorn
